refactor: abstract auth db interactions into AuthRepository
